refactor: merge outlook-mail and outlook-calendar into single outlook plugin
Closes #229
Closes #228

- New unified outlook plugin with all 15 tools (9 mail + 6 calendar)
- Single config, single OAuth credential set
- All tool names unchanged (no breaking changes for callers)
- Removes outlook-mail and outlook-calendar plugin directories
