--- plugins/outlook-calendar/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "outlook-calendar",
-  "version": "1.0.1",
-  "type": "module",
-  "main": "dist/index.js",
-  "bin": {
-    "outlook-calendar": "./dist/bin/outlook-calendar.js"
-  },
-  "scripts": {
-    "build": "tsup && carapace-generate-cli --entry ./dist/index.js --out ./dist/bin",
-    "test": "vitest run"
-  },
-  "devDependencies": {
-    "@types/node": "^26.1.1",
-    "carapace-plugin-sdk": "^1.0.3",
-    "tsup": "^8.3.0",
-    "typescript": "^7",
-    "vitest": "^4.1.10"
-  },
-  "openclaw": {
-    "extensions": [
-      "./dist/adapter.js"
-    ]
-  },
-  "peerDependencies": {
-    "openclaw": "^2026.4.23"
-  },
-  "peerDependenciesMeta": {
-    "openclaw": {
-      "optional": true
-    }
-  },
-  "dependencies": {
-    "@sinclair/typebox": "^0.34.52"
-  }
-}
