--- plugins/outlook-mail/src/adapter.ts
@@ -1,10 +1,0 @@
-import { createRequire } from "node:module";
-import { createEntry } from "./index.js";
-const require = createRequire(import.meta.url);
-let pluginEntry: unknown;
-try {
-  const sdk = require("openclaw/plugin-sdk/plugin-entry") as { definePluginEntry?: (e: unknown) => unknown };
-  if (typeof sdk.definePluginEntry !== "function") throw new Error("SDK missing definePluginEntry");
-  pluginEntry = sdk.definePluginEntry(createEntry());
-} catch { pluginEntry = createEntry(); }
-export default pluginEntry;
