--- plugins/outlook-mail/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src"
-  },
-  "exclude": ["tests", "vitest.config.ts", "dist", "node_modules"]
-}
