--- plugins/outlook-mail/tsup.config.ts
@@ -1,1 +1,0 @@
-export { default } from "../tsup.base";
