--- plugins/outlook-mail/vitest.config.ts
@@ -1,1 +1,0 @@
-export { default } from "../vitest.base";
