+++ plugins/outlook/src/adapter.ts
