+++ plugins/outlook/tsconfig.json
